+++ es/resource_elasticsearch_opendistro_roles_mapping.go
@@ -241,7 +241,7 @@ func resourceElasticsearchPutOpenDistroRolesMapping(d *schema.ResourceData, m in
 			// see https://github.com/opendistro-for-
 			// elasticsearch/security/issues/1095, this should return a 409, but
 			// retry on the 500 as well. We can't parse the message to only retry on
-			// the conlict exception becaues the elastic client doesn't directly
+			// the conflict exception because the elastic client doesn't directly
 			// expose the error response body
 			RetryStatusCodes: []int{http.StatusConflict, http.StatusInternalServerError},
 			Retrier: elastic7.NewBackoffRetrier(