+++ es/resource_elasticsearch_opendistro_user.go
@@ -234,7 +234,7 @@ func resourceElasticsearchPutOpenDistroUser(d *schema.ResourceData, m interface{
 			// see https://github.com/opendistro-for-
 			// elasticsearch/security/issues/1095, this should return a 409, but
 			// retry on the 500 as well. We can't parse the message to only retry on
-			// the conlict exception becaues the elastic client doesn't directly
+			// the conflict exception because the elastic client doesn't directly
 			// expose the error response body
 			RetryStatusCodes: []int{http.StatusConflict, http.StatusInternalServerError},
 			Retrier: elastic7.NewBackoffRetrier(