+++ examples/resources/elasticsearch_opensearch_ism_policy_mapping/import.sh
@@ -1,2 +1,2 @@
-# Import by poilcy_id
+# Import by policy_id
 terraform import elasticsearch_opensearch_ism_policy_mapping.test policy_1